moved around files, renamed files, removed console logs
