+++ src/ProtectedRoutes.js
@@ -5,7 +5,7 @@ import Navbar from './components/Common/Navbar';
 import ScrollToTop from './components/Utilities/ScrollToTop';
 import DatastoreProvider from './context/DatastoreProvider';
 import Dashboard from './layouts/Dashboard';
-import MemberReportContainer from './layouts/MemberReportContainer';
+import MemberReport from './layouts/MemberReport';
 import Reports from './layouts/Reports';
 import NotFound from './notFound';
 
@@ -16,7 +16,7 @@ export default function ProtectedRoute({ authenticated }) {
       <ScrollToTop />
       <Switch>
         <Route path="/reports" component={Reports} />
-        <Route path="/member/:id" render={({ match }) => <MemberReportContainer id={match.params.id} />} />
+        <Route path="/member/:id" render={({ match }) => <MemberReport id={match.params.id} />} />
         <Route exact path="/:measure/members" component={Dashboard} loggedIn={authenticated} />
         <Route exact path="/:measure" component={Dashboard} loggedIn={authenticated} />
         <Route exact path="/" component={Dashboard} loggedIn={authenticated} />
