+++ src/components/Common/Controller.js
@@ -14,9 +14,13 @@ export async function memberInfoFetch(url, id) {
 
 // D3Container.js
 export async function measureDataFetch(measure) {
+  try {
   const memberUrl = new URL(`${env.REACT_APP_HEDIS_MEASURE_API_URL}members?measurementType=${measure}`)
   const defaultVal = await axios.get(memberUrl).then((values) => values.data)
   return defaultVal
+  } catch (error) {
+    return error
+  }
 }
 
 // App.js
@@ -28,6 +32,6 @@ export async function validateAccessToken(accessToken) {
     }
   } catch (error) {
     localStorage.removeItem('token')
+    return false
   }
-  return false
 }
